refactor: split impls into separate files
